+++ apps/api/src/domains/auth/auth.controller.spec.ts
@@ -17,7 +17,7 @@ import { faker } from '@faker-js/faker';
 import { Test } from '@nestjs/testing';
 import dayjs from 'dayjs';
 
-import { getMockProvider } from '@/utils/test-utils';
+import { getMockProvider } from '@/test-utils/util-functions';
 import { TenantService } from '../tenant/tenant.service';
 import { UserDto } from '../user/dtos';
 import { AuthController } from './auth.controller';
