--- apps/api/src/domains/channel/channel/channel.service.providers.ts
@@ -1,40 +1,0 @@
-/**
- * Copyright 2023 LINE Corporation
- *
- * LINE Corporation licenses this file to you under the Apache License,
- * version 2.0 (the "License"); you may not use this file except in compliance
- * with the License. You may obtain a copy of the License at:
- *
- *   https://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
- * WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the
- * License for the specific language governing permissions and limitations
- * under the License.
- */
-import { getRepositoryToken } from '@nestjs/typeorm';
-
-import { OpensearchRepository } from '@/common/repositories';
-import { ProjectServiceProviders } from '@/domains/project/project/project.service.providers';
-import {
-  getMockProvider,
-  MockOpensearchRepository,
-  mockRepository,
-} from '@/utils/test-utils';
-import { FieldServiceProviders } from '../field/field.service.providers';
-import { ChannelEntity } from './channel.entity';
-import { ChannelMySQLService } from './channel.mysql.service';
-import { ChannelService } from './channel.service';
-
-export const ChannelServiceProviders = [
-  ChannelService,
-  ChannelMySQLService,
-  {
-    provide: getRepositoryToken(ChannelEntity),
-    useValue: mockRepository(),
-  },
-  getMockProvider(OpensearchRepository, MockOpensearchRepository),
-  ...ProjectServiceProviders,
-  ...FieldServiceProviders,
-];
