+++ apps/api/src/domains/channel/channel/channel.service.spec.ts
@@ -18,14 +18,14 @@ import { Test } from '@nestjs/testing';
 import { getRepositoryToken } from '@nestjs/typeorm';
 import type { Repository } from 'typeorm';
 
-import { createFieldDto } from '@/utils/test-util-fixture';
-import { MockOpensearchRepository } from '@/utils/test-utils';
+import { createFieldDto } from '@/test-utils/fixtures';
+import { MockOpensearchRepository } from '@/test-utils/util-functions';
+import { ChannelServiceProviders } from '../../../test-utils/providers/channel.service.providers';
 import { ChannelEntity } from '../../channel/channel/channel.entity';
 import { ProjectEntity } from '../../project/project/project.entity';
 import { FieldEntity } from '../field/field.entity';
 import { ChannelMySQLService } from './channel.mysql.service';
 import { ChannelService } from './channel.service';
-import { ChannelServiceProviders } from './channel.service.providers';
 import { CreateChannelDto, FindByChannelIdDto, UpdateChannelDto } from './dtos';
 import {
   ChannelAlreadyExistsException,
