+++ apps/api/src/domains/channel/option/option.controller.spec.ts
@@ -17,7 +17,7 @@ import { faker } from '@faker-js/faker';
 import { Test } from '@nestjs/testing';
 import { DataSource } from 'typeorm';
 
-import { getMockProvider, MockDataSource } from '@/utils/test-utils';
+import { getMockProvider, MockDataSource } from '@/test-utils/util-functions';
 import { CreateOptionRequestDto } from './dtos/requests';
 import { OptionController } from './option.controller';
 import { OptionEntity } from './option.entity';
