+++ apps/api/src/domains/project/issue/issue.controller.spec.ts
@@ -17,7 +17,7 @@ import { faker } from '@faker-js/faker';
 import { Test } from '@nestjs/testing';
 import { DataSource } from 'typeorm';
 
-import { getMockProvider, MockDataSource } from '@/utils/test-utils';
+import { getMockProvider, MockDataSource } from '@/test-utils/util-functions';
 import {
   CreateIssueRequestDto,
   DeleteIssuesRequestDto,
