+++ apps/api/src/shared/mailing/email-verification-mailing.service.spec.ts
@@ -17,7 +17,7 @@ import { faker } from '@faker-js/faker';
 import { MailerService } from '@nestjs-modules/mailer';
 import { Test } from '@nestjs/testing';
 
-import { getMockProvider } from '@/utils/test-utils';
+import { getMockProvider } from '@/test-utils/util-functions';
 import { EmailVerificationMailingService } from './email-verification-mailing.service';
 
 describe('first', () => {
